backend/requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.15,<3.16
django-cors-headers>=4.6,<4.7
dj-database-url>=2.2,<2.3
whitenoise>=6.8,<6.9
gunicorn>=22.0,<23
psycopg2-binary>=2.9,<3
